adds component for displaying calculator summary, removes commented out code
